refactor(angular): simplify the file structure
