+++ packages/angular/src/providers/popover-controller.ts
@@ -1,12 +1,21 @@
-import { Injectable } from '@angular/core';
-import { PopoverController as PopoverControllerBase } from '@ionic/angular/common';
+import { Injector, inject, EnvironmentInjector } from '@angular/core';
+import { AngularDelegate, OverlayBaseController } from '@ionic/angular/common';
+import type { PopoverOptions } from '@ionic/core';
 import { popoverController } from '@ionic/core';
 
-@Injectable({
-  providedIn: 'root',
-})
-export class PopoverController extends PopoverControllerBase {
+export class PopoverController extends OverlayBaseController<PopoverOptions, HTMLIonPopoverElement> {
+  private angularDelegate = inject(AngularDelegate);
+  private injector = inject(Injector);
+  private environmentInjector = inject(EnvironmentInjector);
+
   constructor() {
     super(popoverController);
+  }
+
+  create(opts: PopoverOptions): Promise<HTMLIonPopoverElement> {
+    return super.create({
+      ...opts,
+      delegate: this.angularDelegate.create(this.environmentInjector, this.injector, 'popover'),
+    });
   }
 }
